backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.99.9
librosa==0.11.0
soundfile==0.13.1
python-dotenv==1.0.0